requirements.txt
pandas 
numpy 
matplotlib
seaborn
scikit-learn
flaskpip
flask_cors
gunicorn
nltk
pickle

-e .